refactor: remove admin folder underscore prefix
